refactor on repository package name and removed unused import
