+++ src/main/java/com/maurosalani/project/attsd/repository/UserRepository.java
@@ -1,4 +1,4 @@
-package com.maurosalani.project.attsd.repositories;
+package com.maurosalani.project.attsd.repository;
 
 import java.util.List;
